% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cypher_clauses.R
\name{LIMIT}
\alias{LIMIT}
\title{LIMIT}
\usage{
LIMIT(lhs, limit)
}
\arguments{
\item{lhs}{the left-hand side of the cypher query. Typically, the output of a pipe %>%}

\item{limit}{the limit threshold}
}
\value{
LIMIT
}
\description{
LIMIT
}
