% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_cypher_clauses.R
\name{node}
\alias{node}
\title{Create a node}
\usage{
node(lhs, ..., labels = NULL)
}
\arguments{
\item{lhs}{the left-hand side of the cypher query. Typically, the output of a pipe %>%}

\item{...}{the label, with an optional name variable}

\item{labels}{the properties of the nodes, passed as a name vector}
}
\value{
(node)
}
\description{
Create a node
}
