% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_cypher_clauses.R
\name{to}
\alias{to}
\alias{from}
\title{Create a relationship}
\usage{
to(lhs, ..., labels = NULL)

from(lhs, ..., labels = NULL)
}
\arguments{
\item{lhs}{the left-hand side of the cypher query. Typically, the output of a pipe %>%}

\item{...}{the label, with an optional name variable}

\item{labels}{the properties of the nodes, passed as a name vector}
}
\value{
\code{-[:relationship]-}
}
\description{
Create a relationship
}
