% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cypher_clauses.R
\name{RETURN}
\alias{RETURN}
\title{RETURN}
\usage{
RETURN(lhs, ...)
}
\arguments{
\item{lhs}{the left-hand side of the cypher query. Typically, the output of a pipe %>%}

\item{...}{what to return}
}
\value{
RETURN
}
\description{
RETURN
}
