% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cypher_clauses.R
\name{SET}
\alias{SET}
\title{SET}
\usage{
SET(lhs, ...)
}
\arguments{
\item{lhs}{the left-hand side of the cypher query. Typically, the output of a pipe %>%}

\item{...}{the label, with an optional name variable}
}
\value{
SET x y
}
\description{
SET
}
