% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cypher_clauses.R
\name{WHERE}
\alias{WHERE}
\title{Create a WHERE clause}
\usage{
WHERE(lhs, rhs)
}
\arguments{
\item{lhs}{the left-hand side of the cypher query. Typically, the output of a pipe %>%}

\item{rhs}{the content of the condition}
}
\value{
WHERE()
}
\description{
Create a WHERE clause
}
