% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cypher_clauses.R
\name{WITH}
\alias{WITH}
\title{WITH}
\usage{
WITH(lhs, ...)
}
\arguments{
\item{lhs}{the left-hand side of the cypher query. Typically, the output of a pipe %>%}

\item{...}{the label, with an optional name variable}
}
\description{
Create a WITH clause
}
